GM-0005 style: 상품 도메인, 전체적인 코드 정리 및 테스트 정리2
