+++ src/main/kotlin/com/guacamole/api/common/extension/LongExtension.kt
@@ -1,0 +1,5 @@
+package com.guacamole.api.common.extension
+
+fun Long.greaterOrEqual(other: Long): Boolean {
+    return this >= other
+}
