+++ src/main/kotlin/com/guacamole/api/product/domain/ProductItemPolicy.kt
@@ -1,0 +1,12 @@
+package com.guacamole.api.product.domain
+
+class ProductItemPolicy(
+    var productItemId: Long,
+    var limitUnit: String,
+    var limitCount: Int,
+    var bannedAt: Boolean,
+    var discountUnit: String,
+    var discountRate: Int,
+    var id: Long? = null
+) {
+}
