--- src/main/kotlin/com/guacamole/api/production/product/domain/ProductItem.kt
@@ -1,11 +1,0 @@
-package com.guacamole.api.production.product.domain
-
-class ProductItem(
-    var product_id: Long,
-    var stock_id: Long,
-    var size_unit: Long,
-    var size_rate: Int,
-    var price: Int,
-    var id: Long? = null
-) {
-}
