--- src/main/kotlin/com/guacamole/api/production/product/domain/store/ProductStore.kt
@@ -1,9 +1,0 @@
-package com.guacamole.api.production.product.domain.store
-
-import com.guacamole.api.production.product.domain.Product
-
-interface ProductStore {
-    fun save(product: Product): Product
-    fun findById(productId: Long): Product
-    fun remove(productId: Long)
-}
