--- src/test/kotlin/com/guacamole/api/production/acceptance/AcceptanceTest.kt
@@ -1,19 +1,0 @@
-package com.guacamole.api.production.acceptance
-
-import io.restassured.RestAssured
-import org.junit.jupiter.api.BeforeEach
-import org.springframework.boot.test.context.SpringBootTest
-import org.springframework.boot.test.web.server.LocalServerPort
-
-
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class AcceptanceTest {
-
-    @LocalServerPort
-    var port = 0
-
-    @BeforeEach
-    fun setUp() {
-        RestAssured.port = port
-    }
-}
